Network.AWS.CodeBuild

Description

AWS CodeBuild

AWS CodeBuild is a fully managed build service in the cloud. AWS CodeBuild compiles your source code, runs unit tests, and produces artifacts that are ready to deploy. AWS CodeBuild eliminates the need to provision, manage, and scale your own build servers. It provides prepackaged build environments for the most popular programming languages and build tools, such as Apach Maven, Gradle, and more. You can also fully customize build environments in AWS CodeBuild to use your own build tools. AWS CodeBuild scales automatically to meet peak build requests, and you pay only for the build time you consume. For more information about AWS CodeBuild, see the AWS CodeBuild User Guide .

AWS CodeBuild supports these operations:

BatchGetProjects : Gets information about one or more build projects. A build project defines how AWS CodeBuild will run a build. This includes information such as where to get the source code to build, the build environment to use, the build commands to run, and where to store the build output. A build environment represents a combination of operating system, programming language runtime, and tools that AWS CodeBuild will use to run a build. Also, you can add tags to build projects to help manage your resources and costs.

CreateProject : Creates a build project.

DeleteProject : Deletes a build project.

ListProjects : Gets a list of build project names, with each build project name representing a single build project.

UpdateProject : Changes the settings of an existing build project.

BatchGetBuilds : Gets information about one or more builds.

ListBuilds : Gets a list of build IDs, with each build ID representing a single build.

ListBuildsForProject : Gets a list of build IDs for the specified build project, with each build ID representing a single build.

StartBuild : Starts running a build.

StopBuild : Attempts to stop running a build.

ListCuratedEnvironmentImages : Gets information about Docker images that are managed by AWS CodeBuild.

Service Configuration

codeBuild :: Service

API version 2016-10-06 of the Amazon CodeBuild SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CodeBuild.

ResourceAlreadyExistsException

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified AWS resource cannot be created, because an AWS resource with the same settings already exists.

AccountLimitExceededException

_AccountLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

An AWS service limit was exceeded for the calling AWS account.

InvalidInputException

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

The input value that was provided is not valid.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified AWS resource cannot be found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListProjects

module Network.AWS.CodeBuild.ListProjects

DeleteProject

module Network.AWS.CodeBuild.DeleteProject

UpdateProject

module Network.AWS.CodeBuild.UpdateProject

ListBuilds

module Network.AWS.CodeBuild.ListBuilds

StartBuild

module Network.AWS.CodeBuild.StartBuild

ListBuildsForProject

module Network.AWS.CodeBuild.ListBuildsForProject

BatchGetProjects

module Network.AWS.CodeBuild.BatchGetProjects

BatchGetBuilds

module Network.AWS.CodeBuild.BatchGetBuilds

ListCuratedEnvironmentImages

module Network.AWS.CodeBuild.ListCuratedEnvironmentImages

StopBuild

module Network.AWS.CodeBuild.StopBuild

CreateProject

module Network.AWS.CodeBuild.CreateProject

Types

ArtifactNamespace

data ArtifactNamespace

Constructors

ANBuildId

ANNone

ArtifactPackaging

data ArtifactPackaging

Constructors

None

Zip

ArtifactsType

data ArtifactsType

Constructors

Codepipeline

NoArtifacts

S3

BuildPhaseType

data BuildPhaseType

Constructors

Build

Completed

DownloadSource

Finalizing

Install

PostBuild

PreBuild

Provisioning

Submitted

UploadArtifacts

ComputeType

data ComputeType

Constructors

BuildGENERAL1Large

BuildGENERAL1Medium

BuildGENERAL1Small

EnvironmentType

data EnvironmentType

Constructors

LinuxContainer

LanguageType

data LanguageType

Constructors

Android

Base

Docker

Dotnet

Golang

Java

NodeJs

Python

Ruby

PlatformType

data PlatformType

Constructors

AmazonLinux

Debian

Ubuntu

ProjectSortByType

data ProjectSortByType

Constructors

CreatedTime

LastModifiedTime

Name

SortOrderType

data SortOrderType

Constructors

Ascending

Descending

SourceAuthType

data SourceAuthType

Constructors

Oauth

SourceType

data SourceType

Constructors

STBitbucket

STCodecommit

STCodepipeline

STGithub

STS3

StatusType

data StatusType

Constructors

Failed

Fault

InProgress

Stopped

Succeeded

TimedOut

Build

data Build

Information about a build.

See: build smart constructor.

build :: Build

Creates a value of Build with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bPhases - Information about all previous build phases that are completed and information about any current build phase that is not yet complete.

bBuildComplete - Whether the build has finished. True if completed; otherwise, false.

bArn - The Amazon Resource Name (ARN) of the build.

bStartTime - When the build process started, expressed in Unix time format.

bArtifacts - Information about the output artifacts for the build.

bEnvironment - Information about the build environment for this build.

bInitiator - The entity that started the build. Valid values include: * If AWS CodePipeline started the build, the pipeline's name (for example, codepipeline/my-demo-pipeline ). * If an AWS Identity and Access Management (IAM) user started the build, the user's name (for example MyUserName ). * If the Jenkins plugin for AWS CodeBuild started the build, the string CodeBuild-Jenkins-Plugin .

bCurrentPhase - The current build phase.

bSourceVersion - Any version identifier for the version of the source code to be built.

bLogs - Information about the build's logs in Amazon CloudWatch Logs.

bEndTime - When the build process ended, expressed in Unix time format.

bProjectName - The name of the build project.

bBuildStatus - The current status of the build. Valid values include: * FAILED : The build failed. * FAULT : The build faulted. * IN_PROGRESS : The build is still in progress. * STOPPED : The build stopped. * SUCCEEDED : The build succeeded. * TIMED_OUT : The build timed out.

bSource - Information about the source code to be built.

bId - The unique ID for the build.

bTimeoutInMinutes - How long, in minutes, for AWS CodeBuild to wait before timing out this build if it does not get marked as completed.

bPhases :: Lens' Build [BuildPhase]

Information about all previous build phases that are completed and information about any current build phase that is not yet complete.

bBuildComplete :: Lens' Build (Maybe Bool)

Whether the build has finished. True if completed; otherwise, false.

bArn :: Lens' Build (Maybe Text)

The Amazon Resource Name (ARN) of the build.

bStartTime :: Lens' Build (Maybe UTCTime)

When the build process started, expressed in Unix time format.

bArtifacts :: Lens' Build (Maybe BuildArtifacts)

Information about the output artifacts for the build.

bEnvironment :: Lens' Build (Maybe ProjectEnvironment)

Information about the build environment for this build.

bInitiator :: Lens' Build (Maybe Text)

The entity that started the build. Valid values include: * If AWS CodePipeline started the build, the pipeline's name (for example, codepipeline/my-demo-pipeline ). * If an AWS Identity and Access Management (IAM) user started the build, the user's name (for example MyUserName ). * If the Jenkins plugin for AWS CodeBuild started the build, the string CodeBuild-Jenkins-Plugin .

bCurrentPhase :: Lens' Build (Maybe Text)

The current build phase.

bSourceVersion :: Lens' Build (Maybe Text)

Any version identifier for the version of the source code to be built.

bLogs :: Lens' Build (Maybe LogsLocation)

Information about the build's logs in Amazon CloudWatch Logs.

bEndTime :: Lens' Build (Maybe UTCTime)

When the build process ended, expressed in Unix time format.

bProjectName :: Lens' Build (Maybe Text)

The name of the build project.

bBuildStatus :: Lens' Build (Maybe StatusType)

The current status of the build. Valid values include: * FAILED : The build failed. * FAULT : The build faulted. * IN_PROGRESS : The build is still in progress. * STOPPED : The build stopped. * SUCCEEDED : The build succeeded. * TIMED_OUT : The build timed out.

bSource :: Lens' Build (Maybe ProjectSource)

Information about the source code to be built.

bId :: Lens' Build (Maybe Text)

The unique ID for the build.

bTimeoutInMinutes :: Lens' Build (Maybe Int)

How long, in minutes, for AWS CodeBuild to wait before timing out this build if it does not get marked as completed.

BuildArtifacts

data BuildArtifacts

Information about build output artifacts.

See: buildArtifacts smart constructor.

buildArtifacts :: BuildArtifacts

Creates a value of BuildArtifacts with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

baLocation - Information about the location of the build artifacts.

baMd5sum - The MD5 hash of the build artifact. You can use this hash along with a checksum tool to confirm both file integrity and authenticity.

baSha256sum - The SHA-256 hash of the build artifact. You can use this hash along with a checksum tool to confirm both file integrity and authenticity.

baLocation :: Lens' BuildArtifacts (Maybe Text)

Information about the location of the build artifacts.

baMd5sum :: Lens' BuildArtifacts (Maybe Text)

The MD5 hash of the build artifact. You can use this hash along with a checksum tool to confirm both file integrity and authenticity.

baSha256sum :: Lens' BuildArtifacts (Maybe Text)

The SHA-256 hash of the build artifact. You can use this hash along with a checksum tool to confirm both file integrity and authenticity.

BuildPhase

data BuildPhase

Information about a stage for a build.

See: buildPhase smart constructor.

buildPhase :: BuildPhase

Creates a value of BuildPhase with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bpContexts - Additional information about a build phase, especially to help troubleshoot a failed build.

bpStartTime - When the build phase started, expressed in Unix time format.

bpPhaseStatus - The current status of the build phase. Valid values include: * FAILED : The build phase failed. * FAULT : The build phase faulted. * IN_PROGRESS : The build phase is still in progress. * STOPPED : The build phase stopped. * SUCCEEDED : The build phase succeeded. * TIMED_OUT : The build phase timed out.

bpPhaseType - The name of the build phase. Valid values include: * BUILD : Core build activities typically occur in this build phase. * COMPLETED : The build has been completed. * DOWNLOAD_SOURCE : Source code is being downloaded in this build phase. * FINALIZING : The build process is completing in this build phase. * INSTALL : Installation activities typically occur in this build phase. * POST_BUILD : Post-build activities typically occur in this build phase. * PRE_BUILD : Pre-build activities typically occur in this build phase. * PROVISIONING : The build environment is being set up. * SUBMITTED : The build has been submitted. * UPLOAD_ARTIFACTS : Build output artifacts are being uploaded to the output location.

bpEndTime - When the build phase ended, expressed in Unix time format.

bpDurationInSeconds - How long, in seconds, between the starting and ending times of the build's phase.

bpContexts :: Lens' BuildPhase [PhaseContext]

Additional information about a build phase, especially to help troubleshoot a failed build.

bpStartTime :: Lens' BuildPhase (Maybe UTCTime)

When the build phase started, expressed in Unix time format.

bpPhaseStatus :: Lens' BuildPhase (Maybe StatusType)

The current status of the build phase. Valid values include: * FAILED : The build phase failed. * FAULT : The build phase faulted. * IN_PROGRESS : The build phase is still in progress. * STOPPED : The build phase stopped. * SUCCEEDED : The build phase succeeded. * TIMED_OUT : The build phase timed out.

bpPhaseType :: Lens' BuildPhase (Maybe BuildPhaseType)

The name of the build phase. Valid values include: * BUILD : Core build activities typically occur in this build phase. * COMPLETED : The build has been completed. * DOWNLOAD_SOURCE : Source code is being downloaded in this build phase. * FINALIZING : The build process is completing in this build phase. * INSTALL : Installation activities typically occur in this build phase. * POST_BUILD : Post-build activities typically occur in this build phase. * PRE_BUILD : Pre-build activities typically occur in this build phase. * PROVISIONING : The build environment is being set up. * SUBMITTED : The build has been submitted. * UPLOAD_ARTIFACTS : Build output artifacts are being uploaded to the output location.

bpEndTime :: Lens' BuildPhase (Maybe UTCTime)

When the build phase ended, expressed in Unix time format.

bpDurationInSeconds :: Lens' BuildPhase (Maybe Integer)

How long, in seconds, between the starting and ending times of the build's phase.

EnvironmentImage

data EnvironmentImage

Information about a Docker image that is managed by AWS CodeBuild.

See: environmentImage smart constructor.

environmentImage :: EnvironmentImage

Creates a value of EnvironmentImage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eiName - The name of the Docker image.

eiDescription - The description of the Docker image.

eiName :: Lens' EnvironmentImage (Maybe Text)

The name of the Docker image.

eiDescription :: Lens' EnvironmentImage (Maybe Text)

The description of the Docker image.

EnvironmentLanguage

data EnvironmentLanguage

A set of Docker images that are related by programming language and are managed by AWS CodeBuild.

See: environmentLanguage smart constructor.

environmentLanguage :: EnvironmentLanguage

Creates a value of EnvironmentLanguage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

elImages - The list of Docker images that are related by the specified programming language.

elLanguage - The programming language for the Docker images.

elImages :: Lens' EnvironmentLanguage [EnvironmentImage]

The list of Docker images that are related by the specified programming language.

elLanguage :: Lens' EnvironmentLanguage (Maybe LanguageType)

The programming language for the Docker images.

EnvironmentPlatform

data EnvironmentPlatform

A set of Docker images that are related by platform and are managed by AWS CodeBuild.

See: environmentPlatform smart constructor.

environmentPlatform :: EnvironmentPlatform

Creates a value of EnvironmentPlatform with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

epPlatform - The platform's name.

epLanguages - The list of programming languages that are available for the specified platform.

epPlatform :: Lens' EnvironmentPlatform (Maybe PlatformType)

The platform's name.

epLanguages :: Lens' EnvironmentPlatform [EnvironmentLanguage]

The list of programming languages that are available for the specified platform.

EnvironmentVariable

data EnvironmentVariable

Information about an environment variable for a build project or a build.

See: environmentVariable smart constructor.

environmentVariable

Arguments

:: Text
evName

-> Text
evValue

-> EnvironmentVariable

Creates a value of EnvironmentVariable with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

evName - The name or key of the environment variable.

evValue - The value of the environment variable. Important: We strongly discourage using environment variables to store sensitive values, especially AWS secret key IDs and secret access keys. Environment variables can be displayed in plain text using tools such as the AWS CodeBuild console and the AWS Command Line Interface (AWS CLI).

evName :: Lens' EnvironmentVariable Text

The name or key of the environment variable.

evValue :: Lens' EnvironmentVariable Text

The value of the environment variable. Important: We strongly discourage using environment variables to store sensitive values, especially AWS secret key IDs and secret access keys. Environment variables can be displayed in plain text using tools such as the AWS CodeBuild console and the AWS Command Line Interface (AWS CLI).

LogsLocation

data LogsLocation

Information about build logs in Amazon CloudWatch Logs.

See: logsLocation smart constructor.

logsLocation :: LogsLocation

Creates a value of LogsLocation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

llDeepLink - The URL to an individual build log in Amazon CloudWatch Logs.

llGroupName - The name of the Amazon CloudWatch Logs group for the build logs.

llStreamName - The name of the Amazon CloudWatch Logs stream for the build logs.

llDeepLink :: Lens' LogsLocation (Maybe Text)

The URL to an individual build log in Amazon CloudWatch Logs.

llGroupName :: Lens' LogsLocation (Maybe Text)

The name of the Amazon CloudWatch Logs group for the build logs.

llStreamName :: Lens' LogsLocation (Maybe Text)

The name of the Amazon CloudWatch Logs stream for the build logs.

PhaseContext

data PhaseContext

Additional information about a build phase that has an error. You can use this information to help troubleshoot a failed build.

See: phaseContext smart constructor.

phaseContext :: PhaseContext

Creates a value of PhaseContext with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pcMessage - An explanation of the build phase's context. This explanation might include a command ID and an exit code.

pcStatusCode - The status code for the context of the build phase.

pcMessage :: Lens' PhaseContext (Maybe Text)

An explanation of the build phase's context. This explanation might include a command ID and an exit code.

pcStatusCode :: Lens' PhaseContext (Maybe Text)

The status code for the context of the build phase.

Project

data Project

Information about a build project.

See: project smart constructor.

project :: Project

Creates a value of Project with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pArn - The Amazon Resource Name (ARN) of the build project.

pArtifacts - Information about the build output artifacts for the build project.

pEnvironment - Information about the build environment for this build project.

pCreated - When the build project was created, expressed in Unix time format.

pName - The name of the build project.

pSource - Information about the build input source code for this build project.

pEncryptionKey - The AWS Key Management Service (AWS KMS) customer master key (CMK) to be used for encrypting the build output artifacts. This is expressed either as the CMK's Amazon Resource Name (ARN) or, if specified, the CMK's alias (using the format alias/alias-name ).

pLastModified - When the build project's settings were last modified, expressed in Unix time format.

pDescription - A description that makes the build project easy to identify.

pServiceRole - The ARN of the AWS Identity and Access Management (IAM) role that enables AWS CodeBuild to interact with dependent AWS services on behalf of the AWS account.

pTags - The tags for this build project. These tags are available for use by AWS services that support AWS CodeBuild build project tags.

pTimeoutInMinutes - How long, in minutes, from 5 to 480 (8 hours), for AWS CodeBuild to wait before timing out any related build that did not get marked as completed. The default is 60 minutes.

pArn :: Lens' Project (Maybe Text)

The Amazon Resource Name (ARN) of the build project.

pArtifacts :: Lens' Project (Maybe ProjectArtifacts)

Information about the build output artifacts for the build project.

pEnvironment :: Lens' Project (Maybe ProjectEnvironment)

Information about the build environment for this build project.

pCreated :: Lens' Project (Maybe UTCTime)

When the build project was created, expressed in Unix time format.

pName :: Lens' Project (Maybe Text)

The name of the build project.

pSource :: Lens' Project (Maybe ProjectSource)

Information about the build input source code for this build project.

pEncryptionKey :: Lens' Project (Maybe Text)

The AWS Key Management Service (AWS KMS) customer master key (CMK) to be used for encrypting the build output artifacts. This is expressed either as the CMK's Amazon Resource Name (ARN) or, if specified, the CMK's alias (using the format alias/alias-name ).

pLastModified :: Lens' Project (Maybe UTCTime)

When the build project's settings were last modified, expressed in Unix time format.

pDescription :: Lens' Project (Maybe Text)

A description that makes the build project easy to identify.

pServiceRole :: Lens' Project (Maybe Text)

The ARN of the AWS Identity and Access Management (IAM) role that enables AWS CodeBuild to interact with dependent AWS services on behalf of the AWS account.

pTags :: Lens' Project [Tag]

The tags for this build project. These tags are available for use by AWS services that support AWS CodeBuild build project tags.

pTimeoutInMinutes :: Lens' Project (Maybe Natural)

How long, in minutes, from 5 to 480 (8 hours), for AWS CodeBuild to wait before timing out any related build that did not get marked as completed. The default is 60 minutes.

ProjectArtifacts

data ProjectArtifacts

Information about the build output artifacts for the build project.

See: projectArtifacts smart constructor.

projectArtifacts

Arguments

:: ArtifactsType
paType

-> ProjectArtifacts

Creates a value of ProjectArtifacts with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

paPackaging - The type of build output artifact to create, as follows: * If type is set to CODEPIPELINE , then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output artifacts instead of AWS CodeBuild. * If type is set to NO_ARTIFACTS , then this value will be ignored if specified, because no build output will be produced. * If type is set to S3 , valid values include: * NONE : AWS CodeBuild will create in the output bucket a folder containing the build output. This is the default if packaging is not specified. * ZIP : AWS CodeBuild will create in the output bucket a ZIP file containing the build output.

paPath - Along with namespaceType and name , the pattern that AWS CodeBuild will use to name and store the output artifact, as follows: * If type is set to CODEPIPELINE , then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output names instead of AWS CodeBuild. * If type is set to NO_ARTIFACTS , then this value will be ignored if specified, because no build output will be produced. * If type is set to S3 , this is the path to the output artifact. If path is not specified, then path will not be used. For example, if path is set to MyArtifacts , namespaceType is set to NONE , and name is set to MyArtifact.zip , then the output artifact would be stored in the output bucket at MyArtifacts/MyArtifact.zip .

paLocation - Information about the build output artifact location, as follows: * If type is set to CODEPIPELINE , then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output locations instead of AWS CodeBuild. * If type is set to NO_ARTIFACTS , then this value will be ignored if specified, because no build output will be produced. * If type is set to S3 , this is the name of the output bucket.

paName - Along with path and namespaceType , the pattern that AWS CodeBuild will use to name and store the output artifact, as follows: * If type is set to CODEPIPELINE , then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output names instead of AWS CodeBuild. * If type is set to NO_ARTIFACTS , then this value will be ignored if specified, because no build output will be produced. * If type is set to S3 , this is the name of the output artifact object. For example, if path is set to MyArtifacts , namespaceType is set to BUILD_ID , and name is set to MyArtifact.zip , then the output artifact would be stored in MyArtifacts/build-ID /MyArtifact.zip .

paNamespaceType - Along with path and name , the pattern that AWS CodeBuild will use to determine the name and location to store the output artifact, as follows: * If type is set to CODEPIPELINE , then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output names instead of AWS CodeBuild. * If type is set to NO_ARTIFACTS , then this value will be ignored if specified, because no build output will be produced. * If type is set to S3 , then valid values include: * BUILD_ID : Include the build ID in the location of the build output artifact. * NONE : Do not include the build ID. This is the default if namespaceType is not specified. For example, if path is set to MyArtifacts , namespaceType is set to BUILD_ID , and name is set to MyArtifact.zip , then the output artifact would be stored in MyArtifacts/build-ID /MyArtifact.zip .

paType - The type of build output artifact. Valid values include: * CODEPIPELINE : The build project will have build output generated through AWS CodePipeline. * NO_ARTIFACTS : The build project will not produce any build output. * S3 : The build project will store build output in Amazon Simple Storage Service (Amazon S3).

paPackaging :: Lens' ProjectArtifacts (Maybe ArtifactPackaging)

The type of build output artifact to create, as follows: * If type is set to CODEPIPELINE , then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output artifacts instead of AWS CodeBuild. * If type is set to NO_ARTIFACTS , then this value will be ignored if specified, because no build output will be produced. * If type is set to S3 , valid values include: * NONE : AWS CodeBuild will create in the output bucket a folder containing the build output. This is the default if packaging is not specified. * ZIP : AWS CodeBuild will create in the output bucket a ZIP file containing the build output.

paPath :: Lens' ProjectArtifacts (Maybe Text)

Along with namespaceType and name , the pattern that AWS CodeBuild will use to name and store the output artifact, as follows: * If type is set to CODEPIPELINE , then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output names instead of AWS CodeBuild. * If type is set to NO_ARTIFACTS , then this value will be ignored if specified, because no build output will be produced. * If type is set to S3 , this is the path to the output artifact. If path is not specified, then path will not be used. For example, if path is set to MyArtifacts , namespaceType is set to NONE , and name is set to MyArtifact.zip , then the output artifact would be stored in the output bucket at MyArtifacts/MyArtifact.zip .

paLocation :: Lens' ProjectArtifacts (Maybe Text)

Information about the build output artifact location, as follows: * If type is set to CODEPIPELINE , then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output locations instead of AWS CodeBuild. * If type is set to NO_ARTIFACTS , then this value will be ignored if specified, because no build output will be produced. * If type is set to S3 , this is the name of the output bucket.

paName :: Lens' ProjectArtifacts (Maybe Text)

Along with path and namespaceType , the pattern that AWS CodeBuild will use to name and store the output artifact, as follows: * If type is set to CODEPIPELINE , then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output names instead of AWS CodeBuild. * If type is set to NO_ARTIFACTS , then this value will be ignored if specified, because no build output will be produced. * If type is set to S3 , this is the name of the output artifact object. For example, if path is set to MyArtifacts , namespaceType is set to BUILD_ID , and name is set to MyArtifact.zip , then the output artifact would be stored in MyArtifacts/build-ID /MyArtifact.zip .

paNamespaceType :: Lens' ProjectArtifacts (Maybe ArtifactNamespace)

Along with path and name , the pattern that AWS CodeBuild will use to determine the name and location to store the output artifact, as follows: * If type is set to CODEPIPELINE , then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output names instead of AWS CodeBuild. * If type is set to NO_ARTIFACTS , then this value will be ignored if specified, because no build output will be produced. * If type is set to S3 , then valid values include: * BUILD_ID : Include the build ID in the location of the build output artifact. * NONE : Do not include the build ID. This is the default if namespaceType is not specified. For example, if path is set to MyArtifacts , namespaceType is set to BUILD_ID , and name is set to MyArtifact.zip , then the output artifact would be stored in MyArtifacts/build-ID /MyArtifact.zip .

paType :: Lens' ProjectArtifacts ArtifactsType

The type of build output artifact. Valid values include: * CODEPIPELINE : The build project will have build output generated through AWS CodePipeline. * NO_ARTIFACTS : The build project will not produce any build output. * S3 : The build project will store build output in Amazon Simple Storage Service (Amazon S3).

ProjectEnvironment

data ProjectEnvironment

Information about the build environment of the build project.

See: projectEnvironment smart constructor.

projectEnvironment

Arguments

:: EnvironmentType
peType

-> Text
peImage

-> ComputeType
peComputeType

-> ProjectEnvironment

Creates a value of ProjectEnvironment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pePrivilegedMode - If set to true, enables running the Docker daemon inside a Docker container; otherwise, false or not specified (the default). This value must be set to true only if this build project will be used to build Docker images, and the specified build environment image is not one provided by AWS CodeBuild with Docker support. Otherwise, all associated builds that attempt to interact with the Docker daemon will fail. Note that you must also start the Docker daemon so that your builds can interact with it as needed. One way to do this is to initialize the Docker daemon in the install phase of your build spec by running the following build commands. (Do not run the following build commands if the specified build environment image is provided by AWS CodeBuild with Docker support.) - nohup usrlocalbindockerd --host=unix://varrundocker.sock --host=tcp:/0.0.0.0:2375 --storage-driver=vfs& - timeout -t 15 sh -c "until docker info; do echo .; sleep 1; done"

peEnvironmentVariables - A set of environment variables to make available to builds for this build project.

peType - The type of build environment to use for related builds.

peImage - The ID of the Docker image to use for this build project.

peComputeType - Information about the compute resources the build project will use. Available values include: * BUILD_GENERAL1_SMALL : Use up to 3 GB memory and 2 vCPUs for builds. * BUILD_GENERAL1_MEDIUM : Use up to 7 GB memory and 4 vCPUs for builds. * BUILD_GENERAL1_LARGE : Use up to 15 GB memory and 8 vCPUs for builds.

pePrivilegedMode :: Lens' ProjectEnvironment (Maybe Bool)

If set to true, enables running the Docker daemon inside a Docker container; otherwise, false or not specified (the default). This value must be set to true only if this build project will be used to build Docker images, and the specified build environment image is not one provided by AWS CodeBuild with Docker support. Otherwise, all associated builds that attempt to interact with the Docker daemon will fail. Note that you must also start the Docker daemon so that your builds can interact with it as needed. One way to do this is to initialize the Docker daemon in the install phase of your build spec by running the following build commands. (Do not run the following build commands if the specified build environment image is provided by AWS CodeBuild with Docker support.) - nohup usrlocalbindockerd --host=unix://varrundocker.sock --host=tcp:/0.0.0.0:2375 --storage-driver=vfs& - timeout -t 15 sh -c "until docker info; do echo .; sleep 1; done"

peEnvironmentVariables :: Lens' ProjectEnvironment [EnvironmentVariable]

A set of environment variables to make available to builds for this build project.

peType :: Lens' ProjectEnvironment EnvironmentType

The type of build environment to use for related builds.

peImage :: Lens' ProjectEnvironment Text

The ID of the Docker image to use for this build project.

peComputeType :: Lens' ProjectEnvironment ComputeType

Information about the compute resources the build project will use. Available values include: * BUILD_GENERAL1_SMALL : Use up to 3 GB memory and 2 vCPUs for builds. * BUILD_GENERAL1_MEDIUM : Use up to 7 GB memory and 4 vCPUs for builds. * BUILD_GENERAL1_LARGE : Use up to 15 GB memory and 8 vCPUs for builds.

ProjectSource

data ProjectSource

Information about the build input source code for the build project.

See: projectSource smart constructor.

projectSource

Arguments

:: SourceType
psType

-> ProjectSource

Creates a value of ProjectSource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

psLocation - Information about the location of the source code to be built. Valid values include: * For source code settings that are specified in the source action of a pipeline in AWS CodePipeline, location should not be specified. If it is specified, AWS CodePipeline will ignore it. This is because AWS CodePipeline uses the settings in a pipeline's source action instead of this value. * For source code in an AWS CodeCommit repository, the HTTPS clone URL to the repository that contains the source code and the build spec (for example, https://git-codecommit./region-ID/ .amazonaws.comv1repos/repo-name ). * For source code in an Amazon Simple Storage Service (Amazon S3) input bucket, the path to the ZIP file that contains the source code (for example, bucket-name /path /to /object-name .zip ) * For source code in a GitHub repository, the HTTPS clone URL to the repository that contains the source and the build spec. Also, you must connect your AWS account to your GitHub account. To do this, use the AWS CodeBuild console to begin creating a build project. When you use the console to connect (or reconnect) with GitHub, on the GitHub Authorize application page that displays, for Organization access , choose Request access next to each repository you want to allow AWS CodeBuild to have access to. Then choose Authorize application . (After you have connected to your GitHub account, you do not need to finish creating the build project, and you may then leave the AWS CodeBuild console.) To instruct AWS CodeBuild to then use this connection, in the source object, set the auth object's type value to OAUTH .

psAuth - Information about the authorization settings for AWS CodeBuild to access the source code to be built. This information is for the AWS CodeBuild console's use only. Your code should not get or set this information directly (unless the build project's source type value is GITHUB ).

psBuildspec - The build spec declaration to use for the builds in this build project. If this value is not specified, a build spec must be included along with the source code to be built.

psType - The type of repository that contains the source code to be built. Valid values include: * CODECOMMIT : The source code is in an AWS CodeCommit repository. * CODEPIPELINE : The source code settings are specified in the source action of a pipeline in AWS CodePipeline. * GITHUB : The source code is in a GitHub repository. * S3 : The source code is in an Amazon Simple Storage Service (Amazon S3) input bucket.

psLocation :: Lens' ProjectSource (Maybe Text)

Information about the location of the source code to be built. Valid values include: * For source code settings that are specified in the source action of a pipeline in AWS CodePipeline, location should not be specified. If it is specified, AWS CodePipeline will ignore it. This is because AWS CodePipeline uses the settings in a pipeline's source action instead of this value. * For source code in an AWS CodeCommit repository, the HTTPS clone URL to the repository that contains the source code and the build spec (for example, https://git-codecommit./region-ID/ .amazonaws.comv1repos/repo-name ). * For source code in an Amazon Simple Storage Service (Amazon S3) input bucket, the path to the ZIP file that contains the source code (for example, bucket-name /path /to /object-name .zip ) * For source code in a GitHub repository, the HTTPS clone URL to the repository that contains the source and the build spec. Also, you must connect your AWS account to your GitHub account. To do this, use the AWS CodeBuild console to begin creating a build project. When you use the console to connect (or reconnect) with GitHub, on the GitHub Authorize application page that displays, for Organization access , choose Request access next to each repository you want to allow AWS CodeBuild to have access to. Then choose Authorize application . (After you have connected to your GitHub account, you do not need to finish creating the build project, and you may then leave the AWS CodeBuild console.) To instruct AWS CodeBuild to then use this connection, in the source object, set the auth object's type value to OAUTH .

psAuth :: Lens' ProjectSource (Maybe SourceAuth)

Information about the authorization settings for AWS CodeBuild to access the source code to be built. This information is for the AWS CodeBuild console's use only. Your code should not get or set this information directly (unless the build project's source type value is GITHUB ).

psBuildspec :: Lens' ProjectSource (Maybe Text)

The build spec declaration to use for the builds in this build project. If this value is not specified, a build spec must be included along with the source code to be built.

psType :: Lens' ProjectSource SourceType

The type of repository that contains the source code to be built. Valid values include: * CODECOMMIT : The source code is in an AWS CodeCommit repository. * CODEPIPELINE : The source code settings are specified in the source action of a pipeline in AWS CodePipeline. * GITHUB : The source code is in a GitHub repository. * S3 : The source code is in an Amazon Simple Storage Service (Amazon S3) input bucket.

SourceAuth

data SourceAuth

Information about the authorization settings for AWS CodeBuild to access the source code to be built.

This information is for the AWS CodeBuild console's use only. Your code should not get or set this information directly (unless the build project's source type value is GITHUB ).

See: sourceAuth smart constructor.

sourceAuth

Arguments

:: SourceAuthType
saType

-> SourceAuth

Creates a value of SourceAuth with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

saResource - The resource value that applies to the specified authorization type.

saType - The authorization type to use. The only valid value is OAUTH , which represents the OAuth authorization type.

saResource :: Lens' SourceAuth (Maybe Text)

The resource value that applies to the specified authorization type.

saType :: Lens' SourceAuth SourceAuthType

The authorization type to use. The only valid value is OAUTH , which represents the OAuth authorization type.

Tag

data Tag

A tag, consisting of a key and a value.

This tag is available for use by AWS services that support tags in AWS CodeBuild.

See: tag smart constructor.

tag :: Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagValue - The tag's value.

tagKey - The tag's key.

tagValue :: Lens' Tag (Maybe Text)

The tag's value.

tagKey :: Lens' Tag (Maybe Text)

The tag's key.