Network.AWS.CodeBuild.BatchGetBuilds

Description

Gets information about builds.

Creating a Request

batchGetBuilds

Arguments

:: NonEmpty Text
bgbIds

-> BatchGetBuilds

Creates a value of BatchGetBuilds with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bgbIds - The IDs of the builds.

data BatchGetBuilds

See: batchGetBuilds smart constructor.

Request Lenses

bgbIds :: Lens' BatchGetBuilds (NonEmpty Text)

The IDs of the builds.

Destructuring the Response

batchGetBuildsResponse

Arguments

:: Int
bgbrsResponseStatus

-> BatchGetBuildsResponse

Creates a value of BatchGetBuildsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bgbrsBuilds - Information about the requested builds.

bgbrsBuildsNotFound - The IDs of builds for which information could not be found.

bgbrsResponseStatus - -- | The response status code.

data BatchGetBuildsResponse

See: batchGetBuildsResponse smart constructor.

Response Lenses

bgbrsBuilds :: Lens' BatchGetBuildsResponse [Build]

Information about the requested builds.

bgbrsBuildsNotFound :: Lens' BatchGetBuildsResponse (Maybe (NonEmpty Text))

The IDs of builds for which information could not be found.

bgbrsResponseStatus :: Lens' BatchGetBuildsResponse Int

- | The response status code.