Network.AWS.CodeBuild.CreateProject

Description

Creates a build project.

Creating a Request

createProject

Arguments

:: Text
cpName

-> ProjectSource
cpSource

-> ProjectArtifacts
cpArtifacts

-> ProjectEnvironment
cpEnvironment

-> CreateProject

Creates a value of CreateProject with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpEncryptionKey - The AWS Key Management Service (AWS KMS) customer master key (CMK) to be used for encrypting the build output artifacts. You can specify either the CMK's Amazon Resource Name (ARN) or, if available, the CMK's alias (using the format alias/alias-name ).

cpDescription - A description that makes the build project easy to identify.

cpServiceRole - The ARN of the AWS Identity and Access Management (IAM) role that enables AWS CodeBuild to interact with dependent AWS services on behalf of the AWS account.

cpTags - A set of tags for this build project. These tags are available for use by AWS services that support AWS CodeBuild build project tags.

cpTimeoutInMinutes - How long, in minutes, from 5 to 480 (8 hours), for AWS CodeBuild to wait until timing out any build that has not been marked as completed. The default is 60 minutes.

cpName - The name of the build project.

cpSource - Information about the build input source code for the build project.

cpArtifacts - Information about the build output artifacts for the build project.

cpEnvironment - Information about the build environment for the build project.

data CreateProject

See: createProject smart constructor.

Request Lenses

cpEncryptionKey :: Lens' CreateProject (Maybe Text)

The AWS Key Management Service (AWS KMS) customer master key (CMK) to be used for encrypting the build output artifacts. You can specify either the CMK's Amazon Resource Name (ARN) or, if available, the CMK's alias (using the format alias/alias-name ).

cpDescription :: Lens' CreateProject (Maybe Text)

A description that makes the build project easy to identify.

cpServiceRole :: Lens' CreateProject (Maybe Text)

The ARN of the AWS Identity and Access Management (IAM) role that enables AWS CodeBuild to interact with dependent AWS services on behalf of the AWS account.

cpTags :: Lens' CreateProject [Tag]

A set of tags for this build project. These tags are available for use by AWS services that support AWS CodeBuild build project tags.

cpTimeoutInMinutes :: Lens' CreateProject (Maybe Natural)

How long, in minutes, from 5 to 480 (8 hours), for AWS CodeBuild to wait until timing out any build that has not been marked as completed. The default is 60 minutes.

cpName :: Lens' CreateProject Text

The name of the build project.

cpSource :: Lens' CreateProject ProjectSource

Information about the build input source code for the build project.

cpArtifacts :: Lens' CreateProject ProjectArtifacts

Information about the build output artifacts for the build project.

cpEnvironment :: Lens' CreateProject ProjectEnvironment

Information about the build environment for the build project.

Destructuring the Response

createProjectResponse

Arguments

:: Int
cprsResponseStatus

-> CreateProjectResponse

Creates a value of CreateProjectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cprsProject - Information about the build project that was created.

cprsResponseStatus - -- | The response status code.

data CreateProjectResponse

See: createProjectResponse smart constructor.

Response Lenses

cprsProject :: Lens' CreateProjectResponse (Maybe Project)

Information about the build project that was created.

cprsResponseStatus :: Lens' CreateProjectResponse Int

- | The response status code.