Network.AWS.CodeBuild.ListBuilds

Description

Gets a list of build IDs, with each build ID representing a single build.

Creating a Request

listBuilds :: ListBuilds

Creates a value of ListBuilds with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbSortOrder - The order to list build IDs. Valid values include: * ASCENDING : List the build IDs in ascending order by build ID. * DESCENDING : List the build IDs in descending order by build ID.

lbNextToken - During a previous call, if there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token . To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

data ListBuilds

See: listBuilds smart constructor.

Request Lenses

lbSortOrder :: Lens' ListBuilds (Maybe SortOrderType)

The order to list build IDs. Valid values include: * ASCENDING : List the build IDs in ascending order by build ID. * DESCENDING : List the build IDs in descending order by build ID.

lbNextToken :: Lens' ListBuilds (Maybe Text)

During a previous call, if there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token . To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

Destructuring the Response

listBuildsResponse

Arguments

:: Int
lbrsResponseStatus

-> ListBuildsResponse

Creates a value of ListBuildsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbrsIds - A list of build IDs, with each build ID representing a single build.

lbrsNextToken - If there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token . To get the next batch of items in the list, call this operation again, adding the next token to the call.

lbrsResponseStatus - -- | The response status code.

data ListBuildsResponse

See: listBuildsResponse smart constructor.

Response Lenses

lbrsIds :: Lens' ListBuildsResponse (Maybe (NonEmpty Text))

A list of build IDs, with each build ID representing a single build.

lbrsNextToken :: Lens' ListBuildsResponse (Maybe Text)

If there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token . To get the next batch of items in the list, call this operation again, adding the next token to the call.

lbrsResponseStatus :: Lens' ListBuildsResponse Int

- | The response status code.