Network.AWS.CodeBuild.ListBuildsForProject

Description

Gets a list of build IDs for the specified build project, with each build ID representing a single build.

Creating a Request

listBuildsForProject

Arguments

:: Text
lbfpProjectName

-> ListBuildsForProject

Creates a value of ListBuildsForProject with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbfpSortOrder - The order to list build IDs. Valid values include: * ASCENDING : List the build IDs in ascending order by build ID. * DESCENDING : List the build IDs in descending order by build ID.

lbfpNextToken - During a previous call, if there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token . To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

lbfpProjectName - The name of the build project.

data ListBuildsForProject

See: listBuildsForProject smart constructor.

Request Lenses

lbfpSortOrder :: Lens' ListBuildsForProject (Maybe SortOrderType)

The order to list build IDs. Valid values include: * ASCENDING : List the build IDs in ascending order by build ID. * DESCENDING : List the build IDs in descending order by build ID.

lbfpNextToken :: Lens' ListBuildsForProject (Maybe Text)

During a previous call, if there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token . To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

lbfpProjectName :: Lens' ListBuildsForProject Text

The name of the build project.

Destructuring the Response

listBuildsForProjectResponse

Arguments

:: Int
lbfprsResponseStatus

-> ListBuildsForProjectResponse

Creates a value of ListBuildsForProjectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbfprsIds - A list of build IDs for the specified build project, with each build ID representing a single build.

lbfprsNextToken - If there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token . To get the next batch of items in the list, call this operation again, adding the next token to the call.

lbfprsResponseStatus - -- | The response status code.

data ListBuildsForProjectResponse

See: listBuildsForProjectResponse smart constructor.

Response Lenses

lbfprsIds :: Lens' ListBuildsForProjectResponse (Maybe (NonEmpty Text))

A list of build IDs for the specified build project, with each build ID representing a single build.

lbfprsNextToken :: Lens' ListBuildsForProjectResponse (Maybe Text)

If there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token . To get the next batch of items in the list, call this operation again, adding the next token to the call.

lbfprsResponseStatus :: Lens' ListBuildsForProjectResponse Int

- | The response status code.