Network.AWS.CodeBuild.ListCuratedEnvironmentImages

Description

Gets information about Docker images that are managed by AWS CodeBuild.

Creating a Request

listCuratedEnvironmentImages :: ListCuratedEnvironmentImages

Creates a value of ListCuratedEnvironmentImages with the minimum fields required to make a request.

data ListCuratedEnvironmentImages

See: listCuratedEnvironmentImages smart constructor.

Destructuring the Response

listCuratedEnvironmentImagesResponse

Arguments

:: Int
lceirsResponseStatus

-> ListCuratedEnvironmentImagesResponse

Creates a value of ListCuratedEnvironmentImagesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lceirsPlatforms - Information about supported platforms for Docker images that are managed by AWS CodeBuild.

lceirsResponseStatus - -- | The response status code.

data ListCuratedEnvironmentImagesResponse

See: listCuratedEnvironmentImagesResponse smart constructor.

Response Lenses

lceirsPlatforms :: Lens' ListCuratedEnvironmentImagesResponse [EnvironmentPlatform]

Information about supported platforms for Docker images that are managed by AWS CodeBuild.

lceirsResponseStatus :: Lens' ListCuratedEnvironmentImagesResponse Int

- | The response status code.