Network.AWS.CodeBuild.ListProjects

Description

Gets a list of build project names, with each build project name representing a single build project.

Creating a Request

listProjects :: ListProjects

Creates a value of ListProjects with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpSortOrder - The order in which to list build projects. Valid values include: * ASCENDING : List the build project names in ascending order. * DESCENDING : List the build project names in descending order. Use sortBy to specify the criterion to be used to list build project names.

lpNextToken - During a previous call, if there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token . To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

lpSortBy - The criterion to be used to list build project names. Valid values include: * CREATED_TIME : List the build project names based on when each build project was created. * LAST_MODIFIED_TIME : List the build project names based on when information about each build project was last changed. * NAME : List the build project names based on each build project's name. Use sortOrder to specify in what order to list the build project names based on the preceding criteria.

data ListProjects

See: listProjects smart constructor.

Request Lenses

lpSortOrder :: Lens' ListProjects (Maybe SortOrderType)

The order in which to list build projects. Valid values include: * ASCENDING : List the build project names in ascending order. * DESCENDING : List the build project names in descending order. Use sortBy to specify the criterion to be used to list build project names.

lpNextToken :: Lens' ListProjects (Maybe Text)

During a previous call, if there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token . To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

lpSortBy :: Lens' ListProjects (Maybe ProjectSortByType)

The criterion to be used to list build project names. Valid values include: * CREATED_TIME : List the build project names based on when each build project was created. * LAST_MODIFIED_TIME : List the build project names based on when information about each build project was last changed. * NAME : List the build project names based on each build project's name. Use sortOrder to specify in what order to list the build project names based on the preceding criteria.

Destructuring the Response

listProjectsResponse

Arguments

:: Int
lprsResponseStatus

-> ListProjectsResponse

Creates a value of ListProjectsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lprsNextToken - If there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token . To get the next batch of items in the list, call this operation again, adding the next token to the call.

lprsProjects - The list of build project names, with each build project name representing a single build project.

lprsResponseStatus - -- | The response status code.

data ListProjectsResponse

See: listProjectsResponse smart constructor.

Response Lenses

lprsNextToken :: Lens' ListProjectsResponse (Maybe Text)

If there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token . To get the next batch of items in the list, call this operation again, adding the next token to the call.

lprsProjects :: Lens' ListProjectsResponse (Maybe (NonEmpty Text))

The list of build project names, with each build project name representing a single build project.

lprsResponseStatus :: Lens' ListProjectsResponse Int

- | The response status code.