Network.AWS.CodeBuild.StartBuild

Description

Starts running a build.

Creating a Request

startBuild

Arguments

:: Text
sbProjectName

-> StartBuild

Creates a value of StartBuild with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sbEnvironmentVariablesOverride - A set of environment variables that overrides, for this build only, the latest ones already defined in the build project.

sbTimeoutInMinutesOverride - The number of build timeout minutes, from 5 to 480 (8 hours), that overrides, for this build only, the latest setting already defined in the build project.

sbSourceVersion - A version of the build input to be built, for this build only. If not specified, the latest version will be used. If specified, must be one of: * For AWS CodeCommit or GitHub: the commit ID to use. * For Amazon Simple Storage Service (Amazon S3): the version ID of the object representing the build input ZIP file to use.

sbBuildspecOverride - A build spec declaration that overrides, for this build only, the latest one already defined in the build project.

sbArtifactsOverride - Build output artifact settings that override, for this build only, the latest ones already defined in the build project.

sbProjectName - The name of the build project to start running a build.

data StartBuild

See: startBuild smart constructor.

Request Lenses

sbEnvironmentVariablesOverride :: Lens' StartBuild [EnvironmentVariable]

A set of environment variables that overrides, for this build only, the latest ones already defined in the build project.

sbTimeoutInMinutesOverride :: Lens' StartBuild (Maybe Natural)

The number of build timeout minutes, from 5 to 480 (8 hours), that overrides, for this build only, the latest setting already defined in the build project.

sbSourceVersion :: Lens' StartBuild (Maybe Text)

A version of the build input to be built, for this build only. If not specified, the latest version will be used. If specified, must be one of: * For AWS CodeCommit or GitHub: the commit ID to use. * For Amazon Simple Storage Service (Amazon S3): the version ID of the object representing the build input ZIP file to use.

sbBuildspecOverride :: Lens' StartBuild (Maybe Text)

A build spec declaration that overrides, for this build only, the latest one already defined in the build project.

sbArtifactsOverride :: Lens' StartBuild (Maybe ProjectArtifacts)

Build output artifact settings that override, for this build only, the latest ones already defined in the build project.

sbProjectName :: Lens' StartBuild Text

The name of the build project to start running a build.

Destructuring the Response

startBuildResponse

Arguments

:: Int
srsResponseStatus

-> StartBuildResponse

Creates a value of StartBuildResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srsBuild - Information about the build to be run.

srsResponseStatus - -- | The response status code.

data StartBuildResponse

See: startBuildResponse smart constructor.

Response Lenses

srsBuild :: Lens' StartBuildResponse (Maybe Build)

Information about the build to be run.

srsResponseStatus :: Lens' StartBuildResponse Int

- | The response status code.