Network.AWS.CodeBuild.UpdateProject

Description

Changes the settings of a build project.

Creating a Request

updateProject

Arguments

:: Text
upName

-> UpdateProject

Creates a value of UpdateProject with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upArtifacts - Information to be changed about the build output artifacts for the build project.

upEnvironment - Information to be changed about the build environment for the build project.

upSource - Information to be changed about the build input source code for the build project.

upEncryptionKey - The replacement AWS Key Management Service (AWS KMS) customer master key (CMK) to be used for encrypting the build output artifacts. You can specify either the CMK's Amazon Resource Name (ARN) or, if available, the CMK's alias (using the format alias/alias-name ).

upDescription - A new or replacement description of the build project.

upServiceRole - The replacement ARN of the AWS Identity and Access Management (IAM) role that enables AWS CodeBuild to interact with dependent AWS services on behalf of the AWS account.

upTags - The replacement set of tags for this build project. These tags are available for use by AWS services that support AWS CodeBuild build project tags.

upTimeoutInMinutes - The replacement value in minutes, from 5 to 480 (8 hours), for AWS CodeBuild to wait before timing out any related build that did not get marked as completed.

upName - The name of the build project.

data UpdateProject

See: updateProject smart constructor.

Request Lenses

upArtifacts :: Lens' UpdateProject (Maybe ProjectArtifacts)

Information to be changed about the build output artifacts for the build project.

upEnvironment :: Lens' UpdateProject (Maybe ProjectEnvironment)

Information to be changed about the build environment for the build project.

upSource :: Lens' UpdateProject (Maybe ProjectSource)

Information to be changed about the build input source code for the build project.

upEncryptionKey :: Lens' UpdateProject (Maybe Text)

The replacement AWS Key Management Service (AWS KMS) customer master key (CMK) to be used for encrypting the build output artifacts. You can specify either the CMK's Amazon Resource Name (ARN) or, if available, the CMK's alias (using the format alias/alias-name ).

upDescription :: Lens' UpdateProject (Maybe Text)

A new or replacement description of the build project.

upServiceRole :: Lens' UpdateProject (Maybe Text)

The replacement ARN of the AWS Identity and Access Management (IAM) role that enables AWS CodeBuild to interact with dependent AWS services on behalf of the AWS account.

upTags :: Lens' UpdateProject [Tag]

The replacement set of tags for this build project. These tags are available for use by AWS services that support AWS CodeBuild build project tags.

upTimeoutInMinutes :: Lens' UpdateProject (Maybe Natural)

The replacement value in minutes, from 5 to 480 (8 hours), for AWS CodeBuild to wait before timing out any related build that did not get marked as completed.

upName :: Lens' UpdateProject Text

The name of the build project.

Destructuring the Response

updateProjectResponse

Arguments

:: Int
uprsResponseStatus

-> UpdateProjectResponse

Creates a value of UpdateProjectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uprsProject - Information about the build project that was changed.

uprsResponseStatus - -- | The response status code.

data UpdateProjectResponse

See: updateProjectResponse smart constructor.

Response Lenses

uprsProject :: Lens' UpdateProjectResponse (Maybe Project)

Information about the build project that was changed.

uprsResponseStatus :: Lens' UpdateProjectResponse Int

- | The response status code.